test(test-optimization): move mocha replay helper tests
